% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoplot-addin.R
\name{gogogplot_addin}
\alias{gogogplot_addin}
\title{gogogplot_addin}
\usage{
gogogplot_addin()
}
\description{
This is a wrapper around \code{gogoplot} to use as an RStudio addin. It
allows you to select an object (e.g. a data-frame) and then run the addin
in RStudio either via menu or a hotkey.
}
\keyword{internal}
