% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gogoplot.R
\docType{package}
\name{gogoplot-package}
\alias{gogoplot-package}
\title{Interactive ggplot2 gadgets.}
\description{
The \pkg{gogoplot} package provides simple HTML gadgets
for interactively building data visualziations with ggplot. You can
then return the plot or the code that built it.
}
\author{
Will Bishop \email{wbishop@tesla.com}
}
