% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoplot.R
\name{gogoplot}
\alias{gogoplot}
\title{Interactively build a plot.}
\usage{
gogoplot(.data, popup = FALSE, width = 900, height = 800)
}
\arguments{
\item{.data}{a data-frame to visualize}

\item{popup}{logical value to display the UI as a pop-up window.
The default is FALSE, which will show the UI in
the Viewer pane.}

\item{width}{width of the pop-up window, in pixels.}

\item{height}{height of the pop-up window, in pixels.}
}
\value{
When using RStudio, the generated plot code is inserted at your
cursor using the \pkg{rstudioapi} \code{\link[rstudioapi]{insertText}}
function. Outside of RStudio, the code is printed as a message.
Plot code is also invisibly returned as a string, if you want
to capture the return object.
}
\description{
Launch an interactive shiny gadget that lets you
build a \pkg{ggplot2} plot from the given data. This function is also
callable using an RStudio Addin -- you can high-light a data-frame
object and select the "gogoplot" addin.
}
\examples{
\dontrun{
library(ggplot2)
library(gogoplot)
gogoplot(mtcars)
}

}
