% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{\%++\%}
\alias{\%++\%}
\title{Add a plot layer and capture plot code}
\usage{
e1 \%++\% e2
}
\arguments{
\item{e1}{an existing ggplot2 plot object}

\item{e2}{a new plot layer to add}
}
\value{
an updated plot object
}
