% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gadget.R
\name{plot_gadget}
\alias{plot_gadget}
\title{Interactively build a plot.}
\usage{
plot_gadget(.data, popup = FALSE)
}
\arguments{
\item{.data}{a data-frame to visualize}

\item{popup}{logical value to display the UI as a pop-up window.
The default is FALSE, which will show the UI in
the Viewer pane.}
}
\value{
the UI supports different return objects:
\itemize{
   \item Return Plot - returns a \pkg{ggplot2} plot object
   \item Return Code - returns a character vector of the plotting code. If
   you're using RStudio, this will insert the code at your cursor using the
   \pkg{rstudioapi} \code{\link[rstudioapi]{insertText}} function.
}
}
\description{
Launch an interactive HTML gadget that lets you
build a \pkg{ggplot2} plot from the given data.
}
\examples{
\dontrun{
library(gogoplot)
plot_gadget(mtcars)
}

}
