% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gogoplot.R
\docType{package}
\name{gogoplot-package}
\alias{gogoplot-package}
\title{Interactive ggplot2 gadgets.}
\description{
The \pkg{gogoplot} package provides a simple HTML gadget
for interactively building data visualziations with \pkg{ggplot2}. This
gadget returns the code used to generate the plot, which makes that plot
reproducible. This is meant as a teaching tool for helping folks get familiar
with \pkg{ggplot2} syntax.
}
\author{
Will Bishop \email{wbishop@tesla.com}
}
