% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoplot.R
\name{gogoplot}
\alias{gogoplot}
\title{Interactively build a plot.}
\usage{
gogoplot(.data, popup = FALSE)
}
\arguments{
\item{.data}{a data-frame to visualize}

\item{popup}{logical value to display the UI as a pop-up window.
The default is FALSE, which will show the UI in
the Viewer pane.}
}
\value{
When using RStudio, the generated plot code is inserted at your
cursor using the \pkg{rstudioapi} \code{\link[rstudioapi]{insertText}}
function. Outside of RStudio, the code is printed as a message.
Plot code is also invisibly returned as a string, if you want to capture
the return object.
}
\description{
Launch an interactive shiny gadget that lets you
build a \pkg{ggplot2} plot from the given data.
}
\examples{
\dontrun{
library(gogoplot)
gogoplot(mtcars)
}

}
